% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LINE.R
\name{rLINE}
\alias{rLINE}
\title{Estimate LINE embeddings for a graph}
\usage{
rLINE(wij, dim = 2, sgd_batches = (length(wij@p) - 1) * 20000, M = 5,
  rho = 1, coords = NULL, verbose = TRUE)
}
\arguments{
\item{wij}{A symmetric sparse matrix of edge weights, in C-compressed format, as created with the \code{Matrix} package.}

\item{dim}{The number of dimensions for the projection space; must be even.}

\item{sgd_batches}{The number of edges to process during SGD; defaults to 20000 * the number of rows in x, as recommended
by the paper authors.}

\item{M}{The number of negative edges to sample for each positive edge.}

\item{rho}{Initial learning rate.}

\item{coords}{An initialized coordinate matrix.}

\item{verbose}{Verbosity}
}
\value{
A dense [N,D] matrix of the LINE projection.
}
\description{
Estimate LINE embeddings for a graph
}
\examples{
\dontrun{
data(wiki)
coords <- LINE(wiki)
coords <- t(coords)
plot(coords, xlim = c(-1.5,1.5), ylim = c(-1.5,1.5))
}
}

